20210317 - 산소포화도 측정 추가, BPM측정후 데이터 삽입, 조회 오류 수정 수면 데이터 안들어가지는 현상 수정

+++ app/src/main/java/sleep/simdori/com/util/ImageProcessing.java
@@ -66,7 +66,7 @@ public abstract class ImageProcessing {
      */
     public static double decodeYUV420SPtoRedBlueGreenAvg(byte[] yuv420sp, int width, int height, int type) {
         if (yuv420sp == null) return 0;
-        final int frameSize = width * height;
+        final double frameSize = width * height;
 
         int sum = decodeYUV420SPtoRedBlueGreenSum(yuv420sp, width, height, type);
